+++ pl_examples/basic_examples/dali_image_classifier.py
