+++ pl_examples/basic_examples/simple_image_classifier.py
